Initial commit of global alignment file.
Used to create alignments of nucleotide sequences.

Determines aligments using a dynamic programming algorithm
resembling the Needleman-Wunsch algorith.

Prints the two original sequences and the global alignment
score to the console, and prints the alignments to a file
called alignments.txt.

Also contains debug code for printing the direction and
cost tables to files, which will be removed after
further testing.